+++ lib/xirr/base.rb
@@ -3,6 +3,7 @@ module Xirr
   #  Base module for XIRR calculation Methods
   module Base
     extend ActiveSupport::Concern
+    require 'inline'
     attr_reader :cf
 
     # @param cf [Cashflow]
@@ -14,17 +15,26 @@ module Xirr
     # Calculates days until last transaction
     # @return [Rational]
     # @param date [Date]
-    def t_in_days(date)
-      (date - cf.min_date) / Xirr::DAYS_IN_YEAR
+    def periods_from_start(date)
+      (date - cf.min_date) / cf.period
     end
 
-    # Net Present Value funtion that will be used to reduce the cashflow
+    # Net Present Value function that will be used to reduce the cashflow
     # @param rate [BigDecimal]
+    # @return [BigDecimal]
     def xnpv(rate)
       cf.inject(0) do |sum, t|
-        sum += t.amount / (1 + rate) ** t_in_days(t.date)
+        sum += xnpv_c rate, t.amount, periods_from_start(t.date)
       end
     end
+
+    inline { |builder|
+      builder.include '<math.h>'
+      builder.c '
+        double xnpv_c(double rate, double amount, double period) {
+          return amount / pow(1 + rate, period);
+        }'
+    }
 
   end
 end