+++ CHANGE_LOG.md
@@ -1,3 +1,37 @@
+## Version 0.5.3
+* Better tests
+* added period option to Xirr
+
+## Version 0.5.2
+* Changed negative limits to return nil
+
+## Version 0.5.1
+* Newton Method will return nil if result is less than 100%
+
+## Version 0.5.0
+* This update will break the old Cashflow initializer
+* Adds named attributes to Cashflow Initializer
+* Allows specific configuration to initializer such as: period, flow (array of transactions)
+* Calling xirr with a guess now requires named attribute
+* If a method is provided, it won't fall back to the secondary method
+
+## Version 0.4.1
+* Added verification to pass on ruby 2.0.0
+
+## Version 0.4.0
+
+* Xirr returns nil and there is now a default settings to replace nil rate.
+* It will compact the flow automatically, unless specified in the defaults.
+* Attention to the new way, the Cashflow is created. Cashflow.new requires a Compacted boolean before the array of flow.
+
+## Version 0.3.1
+
+* Added fallback to secondary calculation method.
+
+## Version 0.3.0
+
+* Moved XNPV function to C.
+
 ## Version 0.2.9
 
 * Cashflow validation excludes zeros